requirements.txt
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
click==8.0.3
fastapi==0.70.1
h11==0.12.0
idna==3.3
iso8601==0.1.16
pydantic==1.8.2
pypika-tortoise==0.1.2
pytz==2021.3
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.18.0
typing_extensions==4.0.1
uvicorn==0.16.0
python-dotenv~=0.19.2